fix: added back bitsandbytes in requirements temporarily; it will be moved to base image
