+++ extensions/business/cybersec/red_mesh/mixins/__init__.py
@@ -2,7 +2,7 @@ from .attestation import _AttestationMixin
 from .risk import _RiskScoringMixin
 from .report import _ReportMixin
 from .live_progress import _LiveProgressMixin
-from .llm_agent import _RedMeshLlmAgentMixin
+from .llm_agent_mixin import _RedMeshLlmAgentMixin
 
 __all__ = [
   "_AttestationMixin",
